backend/requirements.txt
aiohttp==3.8.5
click==8.1.6
annotated-types==0.5.0
fastapi==0.101.1
pydantic<2,>=1
SQLAlchemy<3,>=1.4
starlette==0.27.0
typing==3.7.4.1
mysql-connector-python
typing-extensions==4.6.1
python-multipart==0.0.6
passlib[bcrypt]
python-jose[cryptography]
authlib==1.2.1
uvicorn==0.23.2
httpcore==0.13.3
h11<0.13,>=0.11
itsdangerous==2.1.2
mysqlclient==2.1.1
PyJWT==2.8.0
httpx[http2]
langchain==0.0.257
openai==0.28.0
python-dotenv
cloud-sql-python-connector>=1.4.
pg8000==1.29.8
google-api-python-client==2.75.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.60.0
openapi-schema-pydantic==1.2.4